projects/PROJ-024-bayesian-nonparametrics-for-anomaly-dete/code/requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
pymc>=5.0.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
matplotlib>=3.7.0
pyyaml>=6.0.0

# Testing
pytest>=7.4.0

# Linting and formatting
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0